+++ profit_loss.py
@@ -1,7 +1,7 @@
 # Ques) If the cost price and selling price of an item is input through the keyboard, write a program to determine whether the seller has made profit or incurred loss or no profit no loss. also determine how much profit he made or loss he incurred.
 
-cost_price=int(input("Enter the cost price of the item:"))
-selling_price=int(input("Enter the selling price of the item:"))
+cost_price=float(input("Enter the cost price of the item:"))
+selling_price=float(input("Enter the selling price of the item:"))
 
 # If selling price is more than cost price - Profit.
 if cost_price<selling_price:
